+++ lib/rspec/core/notifications.rb
@@ -153,7 +153,7 @@ module RSpec::Core
       #
       # @return [Array(String)] The example failure message
       def message_lines
-        add_shared_group_line(failure_lines, NullColorizer)
+        add_shared_group_lines(failure_lines, NullColorizer)
       end
 
       # Returns the message generated for this failure colorized line by line.
@@ -161,7 +161,7 @@ module RSpec::Core
       # @param colorizer [#wrap] An object to colorize the message_lines by
       # @return [Array(String)] The example failure message colorized
       def colorized_message_lines(colorizer=::RSpec::Core::Formatters::ConsoleCodes)
-        add_shared_group_line(failure_lines, colorizer).map do |line|
+        add_shared_group_lines(failure_lines, colorizer).map do |line|
           colorizer.wrap line, RSpec.configuration.failure_color
         end
       end
@@ -232,29 +232,12 @@ module RSpec::Core
           end
       end
 
-      def add_shared_group_line(lines, colorizer)
-        unless shared_group_line == ""
-          lines << colorizer.wrap(shared_group_line, RSpec.configuration.default_color)
+      def add_shared_group_lines(lines, colorizer)
+        example.metadata[:shared_group_inclusion_backtrace].each do |frame|
+          lines << colorizer.wrap(frame.description, RSpec.configuration.default_color)
         end
+
         lines
-      end
-
-      def shared_group
-        @shared_group ||= group_and_parent_groups.find { |group| group.metadata[:shared_group_name] }
-      end
-
-      def shared_group_line
-        @shared_group_line ||=
-          if shared_group
-            "Shared Example Group: \"#{shared_group.metadata[:shared_group_name]}\"" \
-             " called from #{backtrace_formatter.backtrace_line(shared_group.location)}"
-          else
-            ""
-          end
-      end
-
-      def group_and_parent_groups
-        example.example_group.parent_groups + [example.example_group]
       end
 
       def read_failed_line